requirements.txt
facenet-pytorch
torch
torchvision
pillow
python-telegram-bot